Update test.js with new spec
